message bubble extracted to new component
